Refactor fixture setup
