+++ e2e/i18n-fixture.ts
@@ -1,24 +1,20 @@
+import test from "@playwright/test";
 import { createI18nFixture } from "playwright-i18next-fixture";
 import en from "translations/en";
 import fi from "translations/fi";
 
-export const i18nFixture = createI18nFixture({
-  // i18n configuration options
-  options: {
-    lng: "en",
-    resources: {
-      en: {
-        translation: en,
-      },
-      fi: {
-        translation: fi,
+export const i18nFixture = test.extend(
+  createI18nFixture({
+    options: {
+      lng: "en",
+      resources: {
+        en: {
+          translation: en,
+        },
+        fi: {
+          translation: fi,
+        },
       },
     },
-  },
-  // Fetch translations in every test or fetch once
-  // Default: true
-  cache: true,
-  // Run as auto fixture to be available through all tests by getI18nInstance()
-  // Default: true
-  auto: true,
-});
+  }),
+);
