+++ e2e/tests/controls.mobile.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from "@playwright/test";
-import { test } from "../playwright.config";
+import { test } from "../fixtures/fixtures";
 
 const dateFormat = "YYYY-MM-DD";
